Added logout in user's icon (temporarly) and renamed the login dir to login-logout
