+++ frontend/src/index.js
@@ -6,7 +6,7 @@ import 'bootstrap/dist/css/bootstrap.css';
 import 'react-bootstrap-table/dist/react-bootstrap-table.min.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import Login from "./user/login/Login";
+import Login from "./user/login-logout/Login";
 import UserList from "./user/UserList";
 //import ProductList from "./product/ProductList"; //NOTE: just to show a list of products
 import ProductGrid from "./product/ProductGrid";
